Change parameter of generated enum fromValue() methods to int
Contrary to what I thought, this has to be int, because the first
argument of FlagSet.fromValues() has to be int - that is the version
that returns an IntValuedEnum.

+++ src/gir2java/GirParser.java
@@ -406,11 +406,8 @@ public class GirParser {
 			
 			//fromValue method
 			JMethod fromValue = enumClass.method(JMod.PUBLIC | JMod.STATIC, concreteIntValuedEnumClass, "fromValue");
-			/* 
-			 * XXX: This parameter is an int in the BridJ type mapping doc. This seems to be incorrect,
-			 * as this method accepts one or more ORed values, and these values are longs. 
-			 */
-			JVar fromValueParam = fromValue.param(long.class, "value");
+			
+			JVar fromValueParam = fromValue.param(int.class, "value");
 			JClass flagSetClass = cm.ref(FlagSet.class);
 			JInvocation fromValueCall = flagSetClass.staticInvoke("fromValue")
 					.arg(fromValueParam)
